requirements.txt
django==4.2.5
django-htmx===1.16.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0